chg: dev: refactor init logging to its own file. !refactor
